refactor(p4-1/8): one isPlainObject/cleanString — twin semantics aligned
Three utils imported the strict toString-based pair from new utils/objects.js;
campaignBrief keeps its dependency-free twin copy (byte parity with
src/lib/campaignBrief.js verified by diff) but both twin files now use the
strict semantic — the typeof-based variant accepted Date/RegExp/Map, a latent
divergence (unreachable via JSON bodies, real for programmatic callers).
